Added some explicit type casts to stop msvc6 from emitting warning messages in csgfx/rgbpixel.h and shadervar.h
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@15869 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csgfx/shadervar.h
@@ -155,7 +155,7 @@ public:
   { 
     Type = INT; 
     Int = value; 
-    float f = value;
+    float f = (float)value;
     VectorValue.Set (f, f, f, f);
     return true; 
   }
